ReadVTK.jl API

ReadVTK.VTKCells — Type

VTKCells{Connectivity, Offsets, Types}

Store the connectivity, offsets, and types information from the VTK file as one-dimensional array-like containers. See the VTK file format documentation for information on how to connect the connectivity and offset arrays to the actual geometric points.

You may use length to determine the number of cells from a VTKCells object.

See also: get_points

ReadVTK.VTKData — Type

VTKData

Convenience type to hold information about available data arrays for cells or points.

Supports a collectible/dictionary-like syntax, e.g., keys(vtk_data) to show available data arrays or vtk_data["varname"] to retrieve the VTKDataArray for variable varname.

ReadVTK.VTKDataArray — Type

VTKDataArray{T, N, Format}

Hold information about a single VTK data array (cell data or point data).

The data type is encoded as T, N represents the size of the second dimension in case of multi-dimensonal arrays (or 1 for a vector), and Format encodes in which format the data is stored in the XML file.

The actual data can be retrieved by calling get_data on the VTKDataArray object.

See also: get_data

ReadVTK.VTKDataArray — Method

VTKDataArray(xml_element, vtk_file)

Create a lightweight container for a given xml_element and the corresponding vtk_file.

The xml_element must be of type DataArray and the vtk_file parameter is required for retrieving the actual data. You can pass a VTKDataArray object to get_data to retrieve the actual data.

See also: get_data

ReadVTK.VTKFile — Type

VTKFile

Hold all relevant information about a VTK XML file that has been read in.

Fields

filename: original path to the VTK file that has been read in

xml_file: object that represents the XML file

file_type: currently only "UnstructuredGrid" or "ImageData" are supported

version: currently only v1.0 is supported

byte_order: can be LittleEndian or BigEndian and must currently be the same as the system's

compressor: can be empty (no compression) or vtkZLibDataCompressor

appended_data: in case of appended data (see XML documentation), the data is stored here for convenient retrieval (otherwise it is empty)

n_points: number of points in the VTK file

n_cells: number of cells in the VTK file`

ReadVTK.VTKFile — Method

VTKFile(filename)

Read in and parse the VTK XML file specified by its filename.

ReadVTK.VTKPrimitives — Type

VTKPrimitives{Connectivity, Offsets}

Store the connectivity and offsets information from the VTK PolyData file as one-dimensional array-like containers. See the VTK file format documentation for information on how to connect the connectivity and offset arrays to the actual geometric points.

You may use length to determine the number of cells from a VTKPrimitives object.

ReadVTK.get_cell_data — Method

get_cell_data(vtk_file)

Retrieve a lightweight VTKData object with the cell data of the given VTK file.

See also: VTKData, get_point_data

ReadVTK.get_cells — Method

get_cells(vtk_file)

Retrieve VTK cell information as an object of type VTKCells.

See also: VTKCells

ReadVTK.get_data — Method

get_data(data_array::VTKDataArray)

Retrieve actual data from a VTKDataArray as a one- or two-dimensional array-like container.

Note: This function is not type stable but could be - help wanted!

ReadVTK.get_example_file — Method

get_example_file(filename; head="main", output_directory=".", force=false)

Retrieve an example file from the ReadVTK_examples repository at commit/branch head and store it in the output_directory. If the file already exists locally, do not download the file again unless force is true. Return the local path to the downloaded file.

ReadVTK.get_imagedata_dataset — Method

get_imagedata_dataset(vtk_file)

Retrieve ImageData dataset from the given VTKFile file.

See also: VTKFile

ReadVTK.get_origin — Method

get_origin(vtk_file)

Retrieve the vector of coordinates of the origin of a uniform grid from the given VTKFile file.

See also: VTKFile

ReadVTK.get_point_data — Method

get_point_data(vtk_file)

Retrieve a lightweight VTKData object with the point data of the given VTK file.

See also: VTKData, get_cell_data

ReadVTK.get_points — Method

get_points(vtk_file)

Retrieve VTK points as a two-dimensional array-like container.

The points are stored in dimension × points format. Note that in VTK, points are always stored three-dimensional, even for 1D or 2D files.

See also: get_cells

ReadVTK.get_primitives — Method

get_polydata_primitives(vtk_file, primitive_type::AbstractString)

Retrieve VTK primitives as an object of type VTKPrimitives. Supported values of primitive type are : "Verts", "Lines", or "Polys".

See also: VTKPrimitives

ReadVTK.get_spacing — Method

get_spacing(vtk_file)

Retrieve a vector with the regular increments in each coordinate direction of the uniform grid from the given VTKFile file.

See also: VTKFile

ReadVTK.get_whole_extent — Method

get_whole_extent(vtk_file)

Retrieve a vector with the WholeExtent 6-entry vector from the uniform grid VTKFile file.

See also: VTKFile